+++ tests/ui/apps/Scroller/ScrollerWithEditableSelectItemByLongPress/ScrollerWithEditableSelectItemByLongPress.module.less
