+++ tests/ui/apps/Scroller/ScrollerWithEditableSelectItemByPress/ScrollerWithEditableSelectItemByPress.module.less
